♻️ refactor: split trace hook into hook.go, log.go, spans.go
- hook.go: types, init, constructor, OTel provider setup, lifecycle, session management
- log.go: slog logging for all 5 hook points + summarizeArgs helper
- spans.go: hook interface methods dispatching to log + OTel spans
- config.go: unchanged (env var config)
